+++ guillotina_cms/api/__init__.py
@@ -6,3 +6,7 @@ from . import search  # noqa
 from . import ws_edit  # noqa
 from . import indices  # noqa
 from . import tiles  # noqa
+from . import images  # noqa
+from . import versioning  # noqa
+from . import vocabularies  # noqa
+from . import following  # noqa