+++ guillotina_cms/interfaces/events.py
@@ -1,0 +1,5 @@
+from zope.interface import interfaces
+
+
+class IWorkflowChangedEvent(interfaces.IObjectEvent):
+    """An object workflow has been modified"""