+++ guillotina_cms/subscribers/__init__.py
@@ -1,0 +1,2 @@
+from . import content  # noqa
+from . import edited  # noqa